requirements.txt
spaces
torch
torchvision
accelerate
git+https://github.com/huggingface/diffusers.git
matplotlib
numpy
opencv_python
pillow
PyYAML
scipy
setuptools
scikit-image
tqdm
transformers
fvcore
cloudpickle
omegaconf
pycocotools
av
gradio
peft
huggingface_hub
python-dotenv
cloudinary
fastapi
uvicorn
httpx
pyngrok 
